requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
python-dotenv
psycopg2-binary
pandas
playwright